move singletenantcontroller to kubecontroller package and rename
