+++ cns/kubecontroller/nodenetworkconfig/conversion.go
@@ -1,4 +1,4 @@
-package kubecontroller
+package nodenetworkconfig
 
 import (
 	"net"
