+++ cns/kubecontroller/nodenetworkconfig/reconciler.go
@@ -1,4 +1,4 @@
-package kubecontroller
+package nodenetworkconfig
 
 import (
 	"context"
